refactor: move binmanrelease.go to it's own file
